requirements.txt
draup_packages
streamlit
python-pptx
gTTS
lxml
python-dotenv
imageio-ffmpeg
pywin32; platform_system == "Windows"
pypdf
